backend/requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
# Add the following based on SkinGPT-4 requirements
# torch==2.0.1
# torchvision==0.15.2
# numpy==1.25.2
# Add other dependencies required by SkinGPT-4
